// src/types.ts
export interface PaginationInfo {
  start: number;
  limit: number;
  more_items_in_collection: boolean;
  next_start?: number;
}

export interface PipedriveResponse<T> {
  success: boolean;
  data: T;
  error?: string;
  error_info?: string;
  additional_data?: {
    pagination?: PaginationInfo;
  };
}

export interface SearchItem<T> {
  result_score: number;
  item: T;
}

export interface SearchData<T> {
  items: Array<SearchItem<T>>;
}
// ...
export interface EmailField {
  value: string;
  primary: boolean;
  label: string;
}

export interface PhoneField {
  value: string;
  primary: boolean;
  label: string;
}

export interface Person {
  [key: string]: unknown;
  id: number;
  name: string;
  first_name: string;
  last_name: string;
  email: EmailField[];
  phone: PhoneField[];
  org_id: { value: number; name: string; address?: string } | null;
  owner_id: { id: number; name: string; email: string };
  add_time: string;
  update_time: string;
  open_deals_count: number;
  closed_deals_count: number;
  won_deals_count: number;
  lost_deals_count: number;
  activities_count: number;
  notes_count: number;
  visible_to: string;
}

export interface PersonSearchResult {
  id: number;
  type: string;
  name: string;
  phones: string[];
  emails: string[];
  visible_to: string;
  owner: { id: number };
  organization: { id: number; name: string } | null;
}